separate DataObject validator and schemecheck wip
