+++ retesteth/CMakeLists.txt
@@ -3,6 +3,7 @@ add_executable(${PROJECT_NAME} ${sources})
 
 target_link_libraries(${PROJECT_NAME} PUBLIC Boost::filesystem Boost::program_options Boost::system Boost::thread jsoncpp_lib_static devcore devcrypto cryptopp-static CURL::libcurl)
 target_include_directories(${PROJECT_NAME} PRIVATE "../")
+target_include_directories(${PROJECT_NAME} PRIVATE "../retesteth")
 
 if(HUNTER_ENABLED)
 	# Find Python executable,
