+++ retesteth/RPCSession.cpp
@@ -31,10 +31,11 @@
 #include <mutex>
 #include <csignal>
 
+#include <dataObject/ConvertJsoncpp.h>
+#include <retesteth/EthChecks.h>
+#include <retesteth/ExitHandler.h>
 #include <retesteth/TestHelper.h>
 #include <retesteth/TestOutputHelper.h>
-#include <retesteth/ExitHandler.h>
-#include <retesteth/EthChecks.h>
 
 using namespace std;
 using namespace dev;
@@ -234,13 +235,13 @@ test::scheme_block RPCSession::eth_getBlockByNumber(string const& _blockNumber, 
 {
 	// NOTE: to_string() converts bool to 0 or 1
 	Json::Value response = rpcCall("eth_getBlockByNumber", { quote(_blockNumber), _fullObjects ? "true" : "false" });
-	return test::scheme_block(test::convertJsonCPPtoData(response));
+    return test::scheme_block(dataobject::ConvertJsoncpptoData(response));
 }
 
 test::scheme_transactionReceipt RPCSession::eth_getTransactionReceipt(string const& _transactionHash)
 {
 	Json::Value const result = rpcCall("eth_getTransactionReceipt", { quote(_transactionHash) });
-	return test::scheme_transactionReceipt(test::convertJsonCPPtoData(result));
+    return test::scheme_transactionReceipt(dataobject::ConvertJsoncpptoData(result));
 }
 
 string RPCSession::eth_blockNumber()
